+++ src/main/java/pro/sky/calculator/CalculatorServis.java
@@ -3,17 +3,19 @@ package pro.sky.calculator;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CalculatorServis implements CalculaytorServiceInterface {
+public class CalculatorServis implements CalculaytorServicelmpl {
     @Override
     public String calculator() {
         return "Добро пожаловать в калькулятор";
     }
+
     @Override
     public int plus(int num1, int num2) {
         return num1 + num2;
     }
+
     @Override
-    public  int minus(int num1, int num2) {
+    public int minus(int num1, int num2) {
         return num1 - num2;
     }
