--- src/main/java/pro/sky/calculator/CalculaytorServiceInterface.java
@@ -1,15 +1,0 @@
-package pro.sky.calculator;
-
- public interface CalculaytorServiceInterface {
-
-        String calculator();
-
-        int plus(int num1, int num2);
-
-        int minus(int num1, int num2);
-
-        int multiply(int num1, int num2);
-
-        int divide(int num1, int num2);
- }
-
