+++ src/main/java/pro/sky/calculator/CalculaytorServicelmpl.java
@@ -1,0 +1,15 @@
+package pro.sky.calculator;
+
+public interface CalculaytorServicelmpl {
+
+    String calculator();
+
+    int plus(int num1, int num2);
+
+    int minus(int num1, int num2);
+
+    int multiply(int num1, int num2);
+
+    int divide(int num1, int num2);
+}
+
